backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==41.0.3 